backend/requirements.txt
Flask==3.0.3
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
Flask-Cors==5.0.0
marshmallow==3.22.0
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
requests-mock==1.12.1
google-adk==0.3.0
Deprecated==1.3.1
